Alphabetize! v.1.1.py
v1.1 Fixes an oversight in v.1.0 that caused the Alphabetize! function to attempt to move folders named "#" or "A"-"Z" into themselves, resulting in an error. This prevented users from adding new contents to a target folder's library, that has already been alphabetized, and re-running the alphabetize function.

The function now ignores these folders when moving contents of the target folder allowing users to continually add contents to a target folder's library and running the alphabetize function to merge those new contents with the existing library.